Feature lang stuff and deprecated

+++ Yggdrasil/src/main/java/com/mjolnir/yggdrasil/controllers/api/LanguageController.java
@@ -1,9 +1,10 @@
-package com.mjolnir.yggdrasil.controllers.api;
+package com.mjolnir.yggdrasil.controllers;
 
 import com.mjolnir.yggdrasil.entities.CountryLanguageEntity;
 import com.mjolnir.yggdrasil.entities.CountryLanguageIdEntity;
 import com.mjolnir.yggdrasil.exceptions.InvalidBodyException;
 import com.mjolnir.yggdrasil.exceptions.ResourceNotFoundException;
+import com.mjolnir.yggdrasil.service.MjolnirApiService;
 import com.mjolnir.yggdrasil.service.WorldService;
 import com.mjolnir.yggdrasil.utilities.WebUtils;
 import jakarta.servlet.http.HttpServletRequest;
@@ -13,7 +14,6 @@ import org.springframework.hateoas.Link;
 import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -25,14 +25,20 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 @RestController
 @RequestMapping("/Yggdrasil/languages")
 public class LanguageController {
+    private final MjolnirApiService mjolnirApiService;
     private final WorldService worldService;
 
-    public LanguageController(WorldService worldService) {
+    public LanguageController(MjolnirApiService mjolnirApiService, WorldService worldService) {
+        this.mjolnirApiService = mjolnirApiService;
         this.worldService = worldService;
     }
 
     @PostMapping
-    public ResponseEntity<EntityModel<CountryLanguageEntity>> createLanguage(@RequestBody CountryLanguageEntity language, HttpServletRequest request) {
+    public ResponseEntity<EntityModel<CountryLanguageEntity>> createLanguage(@RequestBody CountryLanguageEntity language, @RequestHeader(name = "MJOLNIR-API-KEY") String apiKey, HttpServletRequest request) {
+        String requestRole = mjolnirApiService.getRoleFromApiKey(apiKey);
+        if (requestRole == null || !requestRole.equals("FULL_ACCESS"))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized user.");
+
         try {
             worldService.createNewCountryLanguage(language.getId().getCountryCode(), language.getLanguage(), language.getIsOfficial(), language.getPercentage());
         } catch (IllegalArgumentException e) {
@@ -48,6 +54,7 @@ public class LanguageController {
             return ResponseEntity.created(location).body(EntityModel.of(lang.get()));
         }
     }
+
 
     @GetMapping("/{countryCode}/{language}")
     public ResponseEntity<EntityModel<CountryLanguageEntity>> getLanguageById(@PathVariable String countryCode, @PathVariable String language, HttpServletRequest request) {
@@ -66,7 +73,6 @@ public class LanguageController {
     }
 
     @GetMapping("/{countryCode}")
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
     public ResponseEntity<CollectionModel<EntityModel<CountryLanguageEntity>>> getLanguageById(@PathVariable String countryCode, HttpServletRequest request) {
         List<CountryLanguageEntity> languages = worldService.getLanguagesByCountryCode(countryCode);
         List<EntityModel<CountryLanguageEntity>> resources = languages.stream()
@@ -87,7 +93,6 @@ public class LanguageController {
     }
 
     @GetMapping
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
     public ResponseEntity<CollectionModel<EntityModel<CountryLanguageEntity>>> getAllLanguages(HttpServletRequest request) {
         List<CountryLanguageEntity> languages = worldService.getAllLanguages();
         List<EntityModel<CountryLanguageEntity>> resources = languages.stream()
@@ -108,14 +113,18 @@ public class LanguageController {
     }
 
     @PutMapping("/{countryCode}/{language}")
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public ResponseEntity<EntityModel<CountryLanguageEntity>> updateLanguage(@PathVariable String countryCode, @PathVariable String language, @RequestBody CountryLanguageEntity updatedLanguage) {
+    public ResponseEntity<EntityModel<CountryLanguageEntity>> updateLanguage(@PathVariable String countryCode, @PathVariable String language, @RequestBody CountryLanguageEntity updatedLanguage, @RequestHeader(name = "MJOLNIR-API-KEY") String apiKey) {
+        String requestRole = mjolnirApiService.getRoleFromApiKey(apiKey);
+        if (requestRole == null || !requestRole.equals("FULL_ACCESS"))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized user.");
+
         CountryLanguageIdEntity primaryKey = new CountryLanguageIdEntity();
         primaryKey.setCountryCode(countryCode);
         primaryKey.setLanguage(language);
         boolean wasUpdateSuccessful = worldService.updateLanguageById(primaryKey, updatedLanguage);
 
         if (wasUpdateSuccessful) {
+
             Optional<CountryLanguageEntity> lang = worldService.getLanguageById(countryCode, language);
 
             if (lang.isEmpty())
@@ -128,8 +137,11 @@ public class LanguageController {
     }
 
     @DeleteMapping("/{countryCode}/{language}")
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public ResponseEntity<Void> deleteLanguage(@PathVariable String countryCode, @PathVariable String language) {
+    public ResponseEntity<Void> deleteLanguage(@PathVariable String countryCode, @PathVariable String language, @RequestHeader(name = "MJOLNIR-API-KEY") String apiKey) {
+        String requestRole = mjolnirApiService.getRoleFromApiKey(apiKey);
+        if (requestRole == null || !requestRole.equals("FULL_ACCESS"))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized user.");
+
         boolean wasDeleteSuccessful = worldService.deleteLanguageByCountryCodeAndLanguage(countryCode, language);
 
         if (wasDeleteSuccessful) {
